+++ layouts/default.vue
@@ -16,7 +16,6 @@
 
 <script lang="ts">
 import { Component, Vue } from 'nuxt-property-decorator'
-import { NotificationProgrammatic as Notification } from 'buefy'
 import { cryptoWaitReady } from '@polkadot/util-crypto'
 import keyring from '@polkadot/ui-keyring'
 import isShareMode from '@/utils/isShareMode'
@@ -64,38 +63,6 @@ export default class Dashboard extends Vue {
 
   public mounted(): void {
     this.mountWasmCrypto()
-    this.checkVersion()
-  }
-
-  private async checkVersion() {
-    //@ts-ignore
-    const workbox = await window.$workbox;
-    if (workbox) {
-      workbox.addEventListener('installed', (event) => {
-        console.log(
-          'App is being served from cache by a service worker.\n' +
-            'For more details, visit https://pwa.nuxtjs.org/'
-        )
-
-        if (event.isUpdate) {
-          console.log('New content is available; please refresh.')
-          const notif = Notification.open({
-            message: 'New version is ready. Close to upgrade.',
-            queue: false,
-            type: 'is-info is-dark',
-            position: 'is-top-left',
-            indefinite: true,
-            hasIcon: true,
-          })
-
-          notif.$on('close', () => {
-            window.sessionStorage.clear()
-            window.localStorage.clear()
-            window.location.reload()
-          })
-        }
-      });
-    }
   }
 
   get isNavbarVisible() {
