+++ nuxt.config.js
@@ -37,6 +37,7 @@ export default {
     '~/plugins/filters',
     '~/plugins/globalVariables',
     '~/plugins/metaInfo',
+    '~/plugins/pwa',
     '~/plugins/vueAudioVisual',
     '~/plugins/vueClipboard',
     '~/plugins/vueSocialSharing',
